Move all `wgpu_renderer` stuff into seperate file
